Move TastyTreeExpr and TreeType out of the library
Now that Expr and Type are not sealed these do not need to be in the library
