+++ library/src/scala/quoted/Expr.scala
@@ -4,7 +4,7 @@ import scala.quoted.show.SyntaxHighlight
 
 /** Quoted expression of type `T`
  *
- *  Restriction: only the QuoteContext implementation is allowed to extend this trait.
+ *  Restriction: only the QuoteContext.tasty.internal implementation is allowed to extend this trait.
  *  Any other implementation will result in an undefined behavior.
  */
 trait Expr[+T] {
